Fix api-database
