+++ packages/api-database/source/repositories/api-node-repository.ts
@@ -1,7 +1,7 @@
 import { ApiNodeRepository, ApiNodeRepositoryExtension, RepositoryDataSource } from "../contracts.js";
 import { ApiNode } from "../models/api-node.js";
 import { ApiNodeFilter } from "../search/filters/index.js";
-import { Criteria, Options, Pagination, ResultsPage, Sorting } from "../search/index.js";
+import { Criteria, Options, Pagination, ResultsPage, Sorting } from "../search/types/index.js";
 import { makeExtendedRepository } from "./repository-extension.js";
 
 export const makeApiNodeRepository = (dataSource: RepositoryDataSource): ApiNodeRepository =>
