+++ packages/api-database/source/repositories/transaction-repository.ts
@@ -9,7 +9,7 @@ import {
 } from "../contracts.js";
 import { Transaction } from "../models/index.js";
 import { TransactionFilter } from "../search/filters/transaction-filter.js";
-import { Criteria, Options, Pagination, ResultsPage, Sorting } from "../search/index.js";
+import { Criteria, Options, Pagination, ResultsPage, Sorting } from "../search/types/index.js";
 import { makeExtendedRepository } from "./repository-extension.js";
 
 export const makeTransactionRepository = (dataSource: RepositoryDataSource): TransactionRepository =>
@@ -45,7 +45,7 @@ export const makeTransactionRepository = (dataSource: RepositoryDataSource): Tra
 			// no days parameter, take the stats from each type for its last 20 txs
 			const result = await this.manager.query<FeeStatistics>(
 				`
-				select 
+				select
 					TRUNC(COALESCE(AVG(txs.gas_price), 0)::numeric) AS "avg",
 					TRUNC(COALESCE(MIN(txs.gas_price), 0)::numeric) AS "min",
 					TRUNC(COALESCE(MAX(txs.gas_price), 0)::numeric) AS "max",
