+++ packages/api-database/source/search/filters/api-node-filter.ts
@@ -1,6 +1,6 @@
 import { ApiNode } from "../../models/index.js";
-import { ApiNodeCriteria, OrApiNodeCriteria } from "../criteria.js";
-import { Expression } from "../expressions.js";
+import { ApiNodeCriteria, OrApiNodeCriteria } from "../types/criteria.js";
+import { Expression } from "../types/expressions.js";
 import { handleAndCriteria, handleComparisonCriteria, handleOrCriteria, optimizeExpression } from "../search.js";
 
 export class ApiNodeFilter {
