+++ packages/api-database/source/search/filters/block-filter.ts
@@ -1,6 +1,6 @@
 import { Block } from "../../models/block.js";
-import { BlockCriteria, OrBlockCriteria } from "../criteria.js";
-import { Expression } from "../expressions.js";
+import { BlockCriteria, OrBlockCriteria } from "../types/criteria.js";
+import { Expression } from "../types/expressions.js";
 import { handleAndCriteria, handleComparisonCriteria, handleOrCriteria, optimizeExpression } from "../search.js";
 
 export class BlockFilter {
