+++ packages/api-database/source/search/filters/delegate-filter.ts
@@ -7,8 +7,8 @@ import {
 	DelegateResourceLastBlock,
 	OrDelegateCriteria,
 	OrNumericCriteria,
-} from "../criteria.js";
-import { Expression, JsonFieldCastType } from "../expressions.js";
+} from "../types/criteria.js";
+import { Expression, JsonFieldCastType } from "../types/expressions.js";
 import { handleAndCriteria, handleComparisonCriteria, handleOrCriteria, optimizeExpression } from "../search.js";
 import { WalletFilter } from "./wallet-filter.js";
