+++ packages/api-database/source/search/filters/peer-filter.ts
@@ -1,6 +1,6 @@
 import { Peer } from "../../models/peer.js";
-import { OrPeerCriteria, PeerCriteria } from "../criteria.js";
-import { Expression } from "../expressions.js";
+import { OrPeerCriteria, PeerCriteria } from "../types/criteria.js";
+import { Expression } from "../types/expressions.js";
 import { handleAndCriteria, handleComparisonCriteria, handleOrCriteria, optimizeExpression } from "../search.js";
 
 export class PeerFilter {
