+++ packages/api-database/source/search/filters/receipt-filter.ts
@@ -1,6 +1,6 @@
 import { Receipt } from "../../models/index.js";
-import { OrReceiptCriteria, ReceiptCriteria } from "../criteria.js";
-import { Expression } from "../expressions.js";
+import { OrReceiptCriteria, ReceiptCriteria } from "../types/criteria.js";
+import { Expression } from "../types/expressions.js";
 import { handleAndCriteria, handleOrCriteria, optimizeExpression } from "../search.js";
 
 export class ReceiptFilter {
