+++ packages/api-database/source/search/filters/transaction-filter.ts
@@ -1,7 +1,7 @@
 import { WalletRepository } from "../../contracts.js";
 import { Transaction } from "../../models/index.js";
-import { EqualCriteria, OrTransactionCriteria, TransactionCriteria } from "../criteria.js";
-import { Expression } from "../expressions.js";
+import { EqualCriteria, OrTransactionCriteria, TransactionCriteria } from "../types/criteria.js";
+import { Expression } from "../types/expressions.js";
 import { handleAndCriteria, handleComparisonCriteria, handleOrCriteria, optimizeExpression } from "../search.js";
 
 export class TransactionFilter {
