+++ packages/api-database/source/search/filters/wallet-filter.ts
@@ -1,8 +1,8 @@
 import { isObject } from "@mainsail/utils";
 
 import { Wallet } from "../../models/index.js";
-import { EqualCriteria, OrWalletCriteria, WalletCriteria } from "../criteria.js";
-import { Expression, JsonFieldCastType, OrExpression } from "../expressions.js";
+import { EqualCriteria, OrWalletCriteria, WalletCriteria } from "../types/criteria.js";
+import { Expression, JsonFieldCastType, OrExpression } from "../types/expressions.js";
 import { handleAndCriteria, handleComparisonCriteria, handleOrCriteria, optimizeExpression } from "../search.js";
 
 export class WalletFilter {
