+++ packages/api-database/source/search/search.ts
@@ -1,4 +1,4 @@
-import { NumericCriteria, OrCriteria } from "./criteria.js";
+import { NumericCriteria, OrCriteria } from "./types/criteria.js";
 import {
 	AndExpression,
 	BetweenExpression,
@@ -8,7 +8,7 @@ import {
 	JsonFieldAccessor,
 	LessThanEqualExpression,
 	OrExpression,
-} from "./expressions.js";
+} from "./types/expressions.js";
 
 export const optimizeExpression = <TEntity>(expression: Expression<TEntity>): Expression<TEntity> => {
 	switch (expression.op) {
