+++ packages/api-database/source/search/types/index.ts
@@ -1,0 +1,26 @@
+import { JsonFieldAccessor } from "./expressions.js";
+
+export * as Criteria from "./criteria.js";
+export * as Expressions from "./expressions.js";
+
+export type SortFragment = {
+	property: string;
+	direction: "asc" | "desc";
+	jsonFieldAccessor?: JsonFieldAccessor;
+};
+export type Sorting = SortFragment[];
+
+export type Pagination = {
+	offset: number;
+	limit: number;
+};
+
+export type Options = {
+	estimateTotalCount?: boolean;
+};
+
+export type ResultsPage<T> = {
+	results: T[];
+	totalCount: number;
+	meta: { totalCountIsEstimate: boolean };
+};
